jobs: refactor code related to waiting for jobs to new file
Release note: None
