+++ pkg/ccl/backupccl/restore_job.go
@@ -1709,7 +1709,7 @@ func (r *restoreResumer) doResume(ctx context.Context, execCtx interface{}) erro
 			return err
 		}
 
-		p.ExecCfg().JobRegistry.NotifyToAdoptJobs(ctx)
+		p.ExecCfg().JobRegistry.NotifyToAdoptJobs()
 		if fn := r.testingKnobs.afterPublishingDescriptors; fn != nil {
 			if err := fn(); err != nil {
 				return err
@@ -1814,7 +1814,7 @@ func (r *restoreResumer) doResume(ctx context.Context, execCtx interface{}) erro
 	}
 	// Reload the details as we may have updated the job.
 	details = r.job.Details().(jobspb.RestoreDetails)
-	p.ExecCfg().JobRegistry.NotifyToAdoptJobs(ctx)
+	p.ExecCfg().JobRegistry.NotifyToAdoptJobs()
 
 	if details.DescriptorCoverage == tree.AllDescriptors {
 		// We restore the system tables from the main data bundle so late because it
