+++ pkg/jobs/registry.go
@@ -11,9 +11,7 @@
 package jobs
 
 import (
-	"bytes"
 	"context"
-	"fmt"
 	"os"
 	"strconv"
 	"strings"
@@ -37,7 +35,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/protoutil"
-	"github.com/cockroachdb/cockroach/pkg/util/retry"
 	"github.com/cockroachdb/cockroach/pkg/util/stop"
 	"github.com/cockroachdb/cockroach/pkg/util/syncutil"
 	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
@@ -261,111 +258,6 @@ func (r *Registry) makeCtx() (context.Context, func()) {
 // MakeJobID generates a new job ID.
 func (r *Registry) MakeJobID() jobspb.JobID {
 	return jobspb.JobID(builtins.GenerateUniqueInt(r.nodeID.SQLInstanceID()))
-}
-
-// NotifyToAdoptJobs notifies the job adoption loop to start claimed jobs.
-func (r *Registry) NotifyToAdoptJobs(context.Context) {
-	select {
-	case r.adoptionCh <- resumeClaimedJobs:
-	default:
-	}
-}
-
-// WaitForJobs waits for a given list of jobs to reach some sort
-// of terminal state.
-func (r *Registry) WaitForJobs(
-	ctx context.Context, ex sqlutil.InternalExecutor, jobs []jobspb.JobID,
-) error {
-	if len(jobs) == 0 {
-		return nil
-	}
-	buf := bytes.Buffer{}
-	for i, id := range jobs {
-		if i > 0 {
-			buf.WriteString(",")
-		}
-		buf.WriteString(fmt.Sprintf(" %d", id))
-	}
-	// Manually retry instead of using SHOW JOBS WHEN COMPLETE so we have greater
-	// control over retries. Also, avoiding SHOW JOBS prevents us from having to
-	// populate the crdb_internal.jobs vtable.
-	query := fmt.Sprintf(
-		`SELECT count(*) FROM system.jobs WHERE id IN (%s)
-       AND (status != $1 AND status != $2 AND status != $3 AND status != $4 AND status != $5)`,
-		buf.String())
-	for r := retry.StartWithCtx(ctx, retry.Options{
-		InitialBackoff: 5 * time.Millisecond,
-		MaxBackoff:     1 * time.Second,
-		Multiplier:     1.5,
-	}); r.Next(); {
-		// We poll the number of queued jobs that aren't finished. As with SHOW JOBS
-		// WHEN COMPLETE, if one of the jobs is missing from the jobs table for
-		// whatever reason, we'll fail later when we try to load the job.
-		row, err := ex.QueryRowEx(
-			ctx,
-			"poll-show-jobs",
-			nil, /* txn */
-			sessiondata.InternalExecutorOverride{User: security.RootUserName()},
-			query,
-			StatusSucceeded,
-			StatusFailed,
-			StatusCanceled,
-			StatusRevertFailed,
-			StatusPaused,
-		)
-		if err != nil {
-			return errors.Wrap(err, "polling for queued jobs to complete")
-		}
-		if row == nil {
-			return errors.New("polling for queued jobs failed")
-		}
-		count := int64(tree.MustBeDInt(row[0]))
-		if log.V(3) {
-			log.Infof(ctx, "waiting for %d queued jobs to complete", count)
-		}
-		if count == 0 {
-			break
-		}
-	}
-	for i, id := range jobs {
-		j, err := r.LoadJob(ctx, id)
-		if err != nil {
-			return errors.WithHint(
-				errors.Wrapf(err, "job %d could not be loaded", jobs[i]),
-				"The job may not have succeeded.")
-		}
-		if j.Payload().FinalResumeError != nil {
-			decodedErr := errors.DecodeError(ctx, *j.Payload().FinalResumeError)
-			return decodedErr
-		}
-		if j.Payload().Error != "" {
-			return errors.Newf("job %d failed with error: %s", jobs[i], j.Payload().Error)
-		}
-		if j.Status() == StatusPaused {
-			if reason := j.Payload().PauseReason; reason != "" {
-				return errors.Newf("job %d was paused before it completed with reason: %s", jobs[i], reason)
-			}
-			return errors.Newf("job %d was paused before it completed", jobs[i])
-		}
-	}
-	return nil
-}
-
-// Run starts previously unstarted jobs from a list of scheduled
-// jobs. Canceling ctx interrupts the waiting but doesn't cancel the jobs.
-func (r *Registry) Run(
-	ctx context.Context, ex sqlutil.InternalExecutor, jobs []jobspb.JobID,
-) error {
-	if len(jobs) == 0 {
-		return nil
-	}
-	log.Infof(ctx, "scheduled jobs %+v", jobs)
-	r.NotifyToAdoptJobs(ctx)
-	err := r.WaitForJobs(ctx, ex, jobs)
-	if err != nil {
-		return err
-	}
-	return nil
 }
 
 // newJob creates a new Job.
